Move domain loading code to the scripts folder.
Move the code that loads the broker's domain values (e.g., object
class, program activity, cgac, sf_133) from the filestreaming folder
to the scripts folder. This is to have it in the same location as
the code that loads the TAS lookup table, which is also a domain
value.
